Refactor project to use c++ modules

+++ Day1/Day1.cpp
@@ -1,6 +1,6 @@
 #include "Day1.hpp"
 
-int64_t calcFuel(int mass)
+int64_t calcFuel(int64_t mass)
 {
 	return (mass / 3) - 2;
 }
