+++ Day1/Day1.hpp
@@ -1,4 +1,4 @@
 #pragma once
 #include<stdint.h>
 
-int64_t calcFuel(int mass);
+int64_t calcFuel(int64_t mass);
